+++ CHANGES.rst
@@ -5,6 +5,9 @@ Version 2.0.2
 
 Unreleased
 
+-   Fix registering a blueprint twice with differing ``name`` not
+    adding handler functions such as ``before_request`` for the new
+    name. :issue:`4131`
 -   Fix type annotation for ``teardown_*`` methods. :issue:`4093`
 -   Fix type annotation for ``before_request`` and ``before_app_request``
     decorators. :issue:`4104`